Changed date achieved data type to long

+++ Assets/src/PlayerRecord.cs
@@ -2,9 +2,9 @@
 {
     public string Name { get; }
     public int CreditEarned { get; }
-    public int DateAchieved { get; }
+    public long DateAchieved { get; }
 
-    public PlayerRecord(string name, int credit, int dateAchieved)
+    public PlayerRecord(string name, int credit, long dateAchieved)
     {
         this.Name = name;
         this.CreditEarned = credit;
